🚚 rename: InformationCard 위치 변경 #58
