+++ src/pages/MyTicket/components/InformationCard/index.tsx
@@ -1,10 +1,10 @@
+import DefaultCard from "@/common/components/atoms/DefaultCard";
+import { InformationCardProp } from "@/common/types/type";
+import useHistoryModal from "@/hooks/useHistoryModal";
+import InformationModal from "@/pages/MyTicket/components/InformationModal";
 import { BsCalendar2EventFill } from "react-icons/bs";
 import { MdLocationOn } from "react-icons/md";
-import { InformationCardProp } from "../../../../common/types/type";
-import InformationModal from "../../../../components/InformationModal";
 import { useNavigate } from "react-router-dom";
-import useHistoryModal from "../../../../hooks/useHistoryModal";
-import DefaultCard from "@/common/components/atoms/DefaultCard";
 
 export default function InformationCard({
   content: { homeTeamName, awayTeamName, gameStartTime, stadiumName, reservationStatus, id },
