+++ src/pages/MyTicket/components/InformationModal/index.tsx
